requirements.txt
asgiref==3.8.1
boto3==1.35.19
botocore==1.35.19
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1.1
djangorestframework==3.15.2
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
moto==5.0.14
pycparser==2.22
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
responses==0.25.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
xmltodict==0.13.0
asgiref==3.8.1
boto3==1.35.19
botocore==1.35.19
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1.1
djangorestframework==3.15.2
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
moto==5.0.14
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
responses==0.25.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
xmltodict==0.13.0
